MatrixPilot_fbw - reverted back to int/long version
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_fbw@2319 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ MatrixPilot/FlightDirector.h
@@ -18,7 +18,7 @@ typedef enum
     FLIGHT_DIRECTOR_MAX
 } FLIGHT_DIRECTOR;
 
-typedef void (*FD_startDirectorFunc)( int16_t ref );
+typedef void (*FD_startDirectorFunc)( int ref );
 typedef void (*FD_closeDirectorFunc)(void);
 typedef void (*FD_runDirector)(void);
